move to thin editor with namespaces
